feat: refactor shipment and notification services, update model imports, and remove unused files
