+++ src/app/models/ShipmentResponse.model.ts
@@ -1,0 +1,12 @@
+import { ShipmentOrder } from "../models/shipmentOrder.model";
+
+export interface ShipmentResponse {
+  status: string;
+  code: number;
+  data: {
+    shipmentOrder: ShipmentOrder[];
+  };
+  message: {
+    text: string;
+  };
+}
