+++ src/app/models/notification.model.ts
@@ -1,0 +1,9 @@
+export type NotificationType = 'success' | 'error' | 'warning' | 'info';
+export interface Notification {
+  id?: string;
+  message: string;
+  type?: NotificationType;
+  duration?: number;
+  read?: boolean;
+  timestamp?: Date;
+}
