+++ src/app/models/shipmentOrder.model.ts
@@ -1,0 +1,9 @@
+import { ShipmentInfo } from './shipmentInfo.model';
+export interface ShipmentOrder {
+  _id: string;
+  cart: any;
+  shipmentInfo: ShipmentInfo | null;
+  status: string;
+  dateOfOrder: string;
+  dateOfDelivery: string;
+}
